refactor: UI 입력 방식 UIInputservice로 분리
